feat(rust): rename vault_worker -> vault, vault -> vault_sync. add mutex support
