+++ implementations/rust/ockam/ockam_vault_sync_core/src/asymmetric_vault.rs
@@ -1,29 +1,38 @@
-use crate::{Vault, VaultRequestMessage, VaultResponseMessage, VaultSyncCoreError};
+use crate::{
+    VaultRequestMessage, VaultResponseMessage, VaultSync, VaultSyncCoreError, VaultSyncState,
+};
 use ockam_core::Result;
 use ockam_node::block_future;
 use ockam_vault_core::{AsymmetricVault, PublicKey, Secret};
 
-impl AsymmetricVault for Vault {
+impl AsymmetricVault for VaultSync {
     fn ec_diffie_hellman(
         &mut self,
         context: &Secret,
         peer_public_key: &PublicKey,
     ) -> Result<Secret> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::EcDiffieHellman {
-                context: context.clone(),
-                peer_public_key: peer_public_key.clone(),
-            })
-            .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::EcDiffieHellman {
+                        context: context.clone(),
+                        peer_public_key: peer_public_key.clone(),
+                    })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::EcDiffieHellman(s) = resp {
-                Ok(s)
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
-            }
-        })
+                if let VaultResponseMessage::EcDiffieHellman(s) = resp {
+                    Ok(s)
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => mutex
+                .lock()
+                .unwrap()
+                .ec_diffie_hellman(context, peer_public_key),
+        }
     }
 }
