+++ implementations/rust/ockam/ockam_vault_sync_core/src/secret_vault.rs
@@ -1,108 +1,136 @@
-use crate::{Vault, VaultRequestMessage, VaultResponseMessage, VaultSyncCoreError};
+use crate::{
+    VaultRequestMessage, VaultResponseMessage, VaultSync, VaultSyncCoreError, VaultSyncState,
+};
 use ockam_core::Result;
 use ockam_node::block_future;
 use ockam_vault_core::{PublicKey, Secret, SecretAttributes, SecretKey, SecretVault};
 
-impl SecretVault for Vault {
+impl SecretVault for VaultSync {
     fn secret_generate(&mut self, attributes: SecretAttributes) -> Result<Secret> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::SecretGenerate { attributes })
-                .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::SecretGenerate { attributes })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::SecretGenerate(s) = resp {
-                Ok(s)
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
-            }
-        })
+                if let VaultResponseMessage::SecretGenerate(s) = resp {
+                    Ok(s)
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => mutex.lock().unwrap().secret_generate(attributes),
+        }
     }
 
     fn secret_import(&mut self, secret: &[u8], attributes: SecretAttributes) -> Result<Secret> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::SecretImport {
-                secret: secret.into(),
-                attributes,
-            })
-            .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::SecretImport {
+                        secret: secret.into(),
+                        attributes,
+                    })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::SecretImport(s) = resp {
-                Ok(s)
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
+                if let VaultResponseMessage::SecretImport(s) = resp {
+                    Ok(s)
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => {
+                mutex.lock().unwrap().secret_import(secret, attributes)
             }
-        })
+        }
     }
 
     fn secret_export(&mut self, context: &Secret) -> Result<SecretKey> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::SecretExport {
-                context: context.clone(),
-            })
-            .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::SecretExport {
+                        context: context.clone(),
+                    })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::SecretExport(s) = resp {
-                Ok(s)
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
-            }
-        })
+                if let VaultResponseMessage::SecretExport(s) = resp {
+                    Ok(s)
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => mutex.lock().unwrap().secret_export(context),
+        }
     }
 
     fn secret_attributes_get(&mut self, context: &Secret) -> Result<SecretAttributes> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::SecretAttributesGet {
-                context: context.clone(),
-            })
-            .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::SecretAttributesGet {
+                        context: context.clone(),
+                    })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::SecretAttributesGet(s) = resp {
-                Ok(s)
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
-            }
-        })
+                if let VaultResponseMessage::SecretAttributesGet(s) = resp {
+                    Ok(s)
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => mutex.lock().unwrap().secret_attributes_get(context),
+        }
     }
 
     fn secret_public_key_get(&mut self, context: &Secret) -> Result<PublicKey> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::SecretPublicKeyGet {
-                context: context.clone(),
-            })
-            .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::SecretPublicKeyGet {
+                        context: context.clone(),
+                    })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::SecretPublicKeyGet(s) = resp {
-                Ok(s)
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
-            }
-        })
+                if let VaultResponseMessage::SecretPublicKeyGet(s) = resp {
+                    Ok(s)
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => mutex.lock().unwrap().secret_public_key_get(context),
+        }
     }
 
     fn secret_destroy(&mut self, context: Secret) -> Result<()> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::SecretDestroy {
-                context: context.clone(),
-            })
-            .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::SecretDestroy {
+                        context: context.clone(),
+                    })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::SecretDestroy = resp {
-                Ok(())
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
-            }
-        })
+                if let VaultResponseMessage::SecretDestroy = resp {
+                    Ok(())
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => mutex.lock().unwrap().secret_destroy(context),
+        }
     }
 }
