+++ implementations/rust/ockam/ockam_vault_sync_core/src/signer.rs
@@ -1,25 +1,31 @@
-use crate::{Vault, VaultRequestMessage, VaultResponseMessage, VaultSyncCoreError};
+use crate::{
+    VaultRequestMessage, VaultResponseMessage, VaultSync, VaultSyncCoreError, VaultSyncState,
+};
 use ockam_core::Result;
 use ockam_node::block_future;
 use ockam_vault_core::{Secret, Signer};
 
-impl Signer for Vault {
+impl Signer for VaultSync {
     fn sign(&mut self, secret_key: &Secret, data: &[u8]) -> Result<[u8; 64]> {
-        block_future(&self.ctx().runtime(), async move {
-            self.send_message(VaultRequestMessage::Sign {
-                secret_key: secret_key.clone(),
-                data: data.into(),
-            })
-            .await?;
+        match &mut self.0 {
+            VaultSyncState::Worker { state } => block_future(&state.ctx().runtime(), async move {
+                state
+                    .send_message(VaultRequestMessage::Sign {
+                        secret_key: secret_key.clone(),
+                        data: data.into(),
+                    })
+                    .await?;
 
-            let resp = self.receive_message().await?;
+                let resp = state.receive_message().await?;
 
-            if let VaultResponseMessage::Sign(s) = resp {
-                Ok(s)
-            } else {
-                Err(VaultSyncCoreError::InvalidResponseType.into())
-            }
-        })
+                if let VaultResponseMessage::Sign(s) = resp {
+                    Ok(s)
+                } else {
+                    Err(VaultSyncCoreError::InvalidResponseType.into())
+                }
+            }),
+            VaultSyncState::Mutex { mutex } => mutex.lock().unwrap().sign(secret_key, data),
+        }
     }
 }
