+++ implementations/rust/ockam/ockam_vault_sync_core/src/vault/response_message.rs
